moved common io layers to a separate folder
