+++ src/main/java/frc/robot/BuildConstants.java
@@ -5,13 +5,13 @@ public final class BuildConstants {
   public static final String MAVEN_GROUP = "";
   public static final String MAVEN_NAME = "2025RobotCode";
   public static final String VERSION = "unspecified";
-  public static final int GIT_REVISION = 25;
-  public static final String GIT_SHA = "942737eea4cfbf98fc8e8b5f699cb025fbd75199";
-  public static final String GIT_DATE = "2025-01-11 13:52:19 EST";
+  public static final int GIT_REVISION = 26;
+  public static final String GIT_SHA = "37009566b168a9e6c99e1d04ad86770ad02c4348";
+  public static final String GIT_DATE = "2025-01-13 19:25:11 EST";
   public static final String GIT_BRANCH = "Coral-scoring-mechanismV-";
-  public static final String BUILD_DATE = "2025-01-13 19:22:04 EST";
-  public static final long BUILD_UNIX_TIME = 1736814124006L;
-  public static final int DIRTY = 1;
+  public static final String BUILD_DATE = "2025-01-14 17:53:12 EST";
+  public static final long BUILD_UNIX_TIME = 1736895192059L;
+  public static final int DIRTY = 128;
 
   private BuildConstants() {}
 }
