+++ src/main/java/frc/robot/subsystems/arms/ArmIOSim.java
@@ -10,6 +10,7 @@ import edu.wpi.first.math.system.plant.DCMotor;
 import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj.simulation.SingleJointedArmSim;
 import frc.robot.constants.SubsystemConstants;
+import frc.robot.subsystems.commoniolayers.ArmIO;
 
 /** Add your docs here. */
 public class ArmIOSim implements ArmIO {
@@ -72,6 +73,7 @@ public class ArmIOSim implements ArmIO {
   }
 
   @Override
+  // runs the motion profile but doesn't generate it
   public void setPositionSetpointDegs(double positionDegs, double ffVolts) {
     appliedVolts = ffVolts;
     pid.setSetpoint(Math.toRadians(positionDegs));
