+++ src/main/java/frc/robot/subsystems/commoniolayers/ArmIO.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.arms;
+package frc.robot.subsystems.commoniolayers;
 
 import org.littletonrobotics.junction.AutoLog;
