+++ src/main/java/frc/robot/subsystems/commoniolayers/FlywheelIO.java
@@ -11,7 +11,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
 
-package frc.robot.subsystems.flywheel;
+package frc.robot.subsystems.commoniolayers;
 
 import org.littletonrobotics.junction.AutoLog;
