+++ src/main/java/frc/robot/subsystems/coralscorer/CoralScorerArmIOSim.java
@@ -1,4 +1,4 @@
-package frc.robot.TSubsystems.CoralScorer;
+package frc.robot.subsystems.coralscorer;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.controller.PIDController;
@@ -6,7 +6,7 @@ import edu.wpi.first.math.system.plant.DCMotor;
 import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj.simulation.SingleJointedArmSim;
 import frc.robot.constants.SubsystemConstants;
-import frc.robot.subsystems.arms.ArmIO;
+import frc.robot.subsystems.commoniolayers.ArmIO;
 
 /** Add your docs here. */
 public class CoralScorerArmIOSim implements ArmIO {
