+++ src/main/java/frc/robot/subsystems/coralscorer/CoralScorerFlywheelIOSim.java
@@ -1,4 +1,4 @@
-package frc.robot.TSubsystems.CoralScorer;
+package frc.robot.subsystems.coralscorer;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.controller.PIDController;
@@ -6,7 +6,7 @@ import edu.wpi.first.math.system.plant.DCMotor;
 import edu.wpi.first.math.system.plant.LinearSystemId;
 import edu.wpi.first.wpilibj.simulation.DCMotorSim;
 import frc.robot.constants.SubsystemConstants;
-import frc.robot.subsystems.flywheel.FlywheelIO;
+import frc.robot.subsystems.commoniolayers.FlywheelIO;
 
 public class CoralScorerFlywheelIOSim implements FlywheelIO {
   // CHANGE THESE VALUES TO MATCH YOUR MOTOR AND GEARBOX
