+++ src/main/java/frc/robot/subsystems/flywheel/Flywheel.java
@@ -22,6 +22,9 @@ import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
 import frc.robot.constants.SimConstants;
+import frc.robot.subsystems.commoniolayers.FlywheelIO;
+import frc.robot.subsystems.commoniolayers.FlywheelIOInputsAutoLogged;
+
 import org.littletonrobotics.junction.AutoLogOutput;
 import org.littletonrobotics.junction.Logger;
